refactor: electron 非依存の純粋モジュールを src/shared/ へ移動
Phase 2 のディレクトリ再編。apmPath / compareVersion / resolvePath /
integrity / getHash とそのテストを src/shared/ に移し、lib/ は
electron 依存モジュール置き場として責務を明確化する。
中身の変更はなし(git mv + import パス更新のみ)。
