+++ src/shared/getHash.ts
