+++ src/shared/integrity.ts
