+++ src/shared/resolvePath.test.ts
